requirements.txt
asgiref==3.2.3
astroid==2.3.1
autopep8==1.4.4
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.5
django-appconf==1.0.3
django-geoip==0.5.2.1
geoip2==2.9.0
httplib2==0.17.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
maxminddb==1.5.2
mccabe==0.6.1
oauth2==1.9.0.post1
oauthlib==3.1.0
pep8==1.7.1
Pillow==6.2.0
pkg-resources==0.0.0
psycopg2==2.8.3
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.2
python-openid==2.2.5
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.3.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.8
wrapt==1.11.2
